Added more Panel separation for better encapsulation.
Added the ClientUtils class which contains methods to launch/help the
GUI. Still working on focus/view transfers.
