--- Source_Code/Client/App/views/LobbyStatus.java
@@ -1,8 +1,0 @@
-package Client.App.views;
-
-import javax.swing.JPanel;
-
-public class LobbyStatus extends JPanel
-{
-
-}
